+++ crates/db/sqlite/src/tests/mod.rs
@@ -11,6 +11,7 @@
 #![allow(clippy::arithmetic_side_effects, clippy::indexing_slicing)]
 
 mod ambiguous;
+mod attach;
 mod auth;
 mod balance;
 mod bytes;